requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
openai